+++ Changelog.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## [0.2.2] - 2019/02/27
+
+### Added
+- Better documentation and examples of API methods
+
+### Fixed
+- Fixed shadowing of built-in `all` function by translateChr()
+
+## [0.2.1] - 2019/02/25
+
+### Added
+- Added support for strandedness in tx2genome()
+
+### Fixed
+- Fixed bug in tx2genome() strand handling
+
 ## [0.2.0] - 2019/01/19
 
 ### Fixed